swifttrack-middleware/services/api-gateway/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pika==1.3.2
psycopg[binary]==3.2.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4